fix tests failing because of scaled factor

+++ packages/world/test/Equip.t.sol
@@ -195,7 +195,7 @@ contract EquipTest is DustTest {
 
     ObjectTypeId mineObjectTypeId = TerrainLib.getBlockType(mineCoord);
     uint128 expectedMassReductionFromTool = 50;
-    ObjectTypeMetadata.setMass(mineObjectTypeId, uint32(expectedMassReductionFromTool));
+    ObjectTypeMetadata.setMass(mineObjectTypeId, expectedMassReductionFromTool);
     EntityId mineEntityId = ReversePosition.get(mineCoord);
     assertFalse(mineEntityId.exists(), "Mine entity already exists");
     assertInventoryHasObject(aliceEntityId, mineObjectTypeId, 0);
